Change behaviour of function code generation to have pops at the end of the function
The code that got removed from the TypeResolver was what was supposed to do this, however
it was based on the incorrect assumption that control would only roll off of the end of the function
if no return statement was present. To fix this, the function code generator just emits extra pops at the
end of the function.