Fix bug that broke basically the entire runtime
So in the function `end_scope` in the Generator class, it emits POP instructions only at the global
scope to clear up the global variables that are created at runtime. Unfortunately, I made it this way
only thinking about functions and how they would work, where the return would clean up the locals and there
would be no reason to emit POPs at the end of the scope. This is completely wrong, since there are other statements,
such as for statements, that introduce a scope themselves. In the case of for statements, the scope is introduced for
loop variables created in the initializer. Since POPs are not emitted at the end of the block, this leads to the issue
where the loop variables are not cleaned up and thus loops inside loops only run once.t