Fix completely broken local variable accessing
Since the TypeResolver would use values.size() to choose the stack slot for variables, it would mean that
instead of a new substack beginning for every function from zero, the stack slot chosen for any variable
access would be as if it was in the global scope. But, because of the way the VM works, it would treat
that global stack slot as the local stack slot for the variable, leading to many nasal demons. The solution
is to make the stack slot for a variable depend on the variable that came before it, and to start the stack slot
for the first parameter of any function from zero.