Fix bug relating to dereferencing
In list indexing and assignment, the index and assigned value are not checked for being
references. Thus, no Instruction::DEREF is emitted. At runtime, this leads to an exception
being thrown for std::bad_varaint_access. The fix is to emit the Instruction::DEREF when needed. for std::bad_variant_access. The fix is to emit the Instruction::DEREF when needed.