Fix segfaults in List expressions
- First one in creation of 2 or more D lists
- Second one in repr() of empty 2 or more D lists

For the first one, when a list of list is initially created, it is essentially initialized
with nullptr for the contained elements, thus trying to assign to those contained lists when
copying or moving causes a segfault.

Same issue with the second one, but instead of assignment its trying to call member functions.