Change the calling convention
Now, the null value pushed as the return value stack slot before calling a function
is included as part of the called function's stack. This means that all accesses to
local variables need to be offset by +1, to account for that null value. As the code
gen for accessing local and global variables share the same call to `emit_stack_slot`,
a null needs to be pushed at the start of each `Chunk` to emulate the same behaviour
for globals as that of the local variables