Change calling convention and implement destructor calls
- The calling convention now includes the NULL return value stack slot
as a stack slot of the called function as well
- Now, `Instruction::RETURN` no longer handles destroying all the locals
on the current stack frame, as this makes calling destructors near impossible
- Instead, returning occurs in 3 steps like so:
  - First, the return value is stored into the NULL stack slot. If a constructor
or destructor is being called, then no store is created because the return value
stack slot contains the object itself
  - Second, all the variables in the stack frame are destroyed. All the destructors
of objects and their subobjects are called here.
  - Third, the actual return occurs, where the `ip` and `chunk` are restored to their
previous values and the stack frame is popped off.