[fix] (RuntimeModule) : add constructors
There is an issue (feature) with `std::vector` where it falls back to the contained type's copy constructor on resizing with `emplace_back` for example if the contained type's move constructor is not noexcept.

This causes an issue where the `std::vector<RuntimeModule>` in `RuntimeManager` has to copy the `RuntimeModule` when resizing, which causes the `std::string`s contained in the `RuntimeModule`'s `Chunk`'s `std::deque<std::string>` to be copied, changing their addresses, which breaks the `HashedString` pointers in the `Value`s contained in the `constants` vector, eventually causing a SIGSEGV for an out of bounds memory read.

To fix this issue, the noexcept move constructor must be defined for `RuntimeModule`