[fix] (TypeResolver/Module) : fix massive bug with `type_scratch_space`
`TypeResolver` has a temporary storage for synthesized types, called `type_scratch_space`. These sythesized types are pointed to by `info` in `synthesized_attrs` for some `Expr`s. When a module is imported, the `Parser` sets up a temporary `CompileManager` to parse/typecheck the module, then extracts the `Module` object from the `CompileManager`, and allows the `CompileManager` to be destroyed. The problem is, some `Expr`s stored inside the `Module` point to types in the `type_scratch_space` of the `TypeResolver` contained in the `CompileManager` which gets destroyed, so they become invalid pointers.

Eventually, this leads to a SIGSEGV, and loads of invalid reads in valgrind. To fix this issue, the `type_scratch_space` is lifted out of `TypeResolver` into `Module` and then `TypeResolver` maintains a pointer to `Module`, which fixes the lifetime issue.